fix(rbac): fix migration script

+++ plugins/rbac-backend/migrations/20240611092136_migrations.js
@@ -5,7 +5,9 @@
 exports.up = async function up(knex) {
   const policyMetadataExist = await knex.schema.hasTable('policy-metadata');
 
-  await knex.schema.dropTable('policy-metadata');
+  if (policyMetadataExist) {
+    await knex.schema.dropTable('policy-metadata');
+  }
 };
 
 /**
